clean up tests for #2341
