+++ facet-json/tests/integration/mod.rs
@@ -1,8 +1,6 @@
 #[path = "backend_historical.rs"]
 pub(crate) mod json_backend;
 
-mod azure_devops_descriptor_map_key_repro;
-mod azure_devops_descriptor_map_key_working_pattern;
 mod flatten_defaults;
 mod flatten_in_externally_tagged_enum;
 mod format_specific_proxy;
@@ -25,6 +23,7 @@ mod issue_2004;
 mod issue_2007;
 mod issue_2010;
 mod issue_2059;
+mod issue_2341_newtype_as_json_object_key;
 mod list_deferred_processing;
 mod metadata_container_flatten_map;
 mod mixed_tagged_untagged;
